requirements.txt
Pillow>=9.0.0
imagehash>=4.3.0
tqdm>=4.64.0
